+++ Assignment 2/openmp_and_mpi.h
@@ -1,7 +1,7 @@
 #ifndef OPENMP_AND_MPI
 #define OPENMP_AND_MPI
 
-void openmp_and_mpi(int improvement, long int n, int p) {
+void openmp_and_mpi(int improvement, long long n, int p) {
 	
 }
